Split keybindings.rs (911 lines) into keybindings/ subdirectory
Split the monolithic keybindings.rs into three focused modules:
- actions.rs (168 lines): BindableAction enum with labels, categories, ALL
- bindings.rs (435 lines): KeyBinding, KeyBindings resource, serialization, plugin
- key_helpers.rs (312 lines): keycode_label, keycode_to_u16, u16_to_keycode, tests

All public items re-exported from mod.rs so external imports are unchanged.
